# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwiftShip Logistics Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /riders:
    get:
      operationId: listRiders
      tags: [riders]
      summary: List all riders
      parameters:
        - name: stationId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rider"
    post:
      operationId: createRider
      tags: [riders]
      summary: Create a rider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRiderBody"
      responses:
        "201":
          description: Rider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
  /riders/{id}:
    get:
      operationId: getRider
      tags: [riders]
      summary: Get a rider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
    patch:
      operationId: updateRider
      tags: [riders]
      summary: Update a rider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRiderBody"
      responses:
        "200":
          description: Updated rider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
    delete:
      operationId: deleteRider
      tags: [riders]
      summary: Delete a rider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
...
  /bank-accounts:
    get:
      operationId: listBankAccounts
      tags: [payments]
      summary: List bank accounts
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [payments]
      summary: Add a bank account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankAccountBody"
      responses:
        "201":
          description: Bank account added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank-accounts/{id}:
    patch:
      operationId: updateBankAccount
      tags: [payments]
      summary: Update a bank account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBankAccountBody"
      responses:
        "200":
          description: Updated bank account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
    delete:
      operationId: deleteBankAccount
      tags: [payments]
      summary: Delete a bank account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Payment Requests
  /payment-requests:
    get:
      operationId: listPaymentRequests
      tags: [payments]
      summary: List payment release requests
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payment requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRequest"
    post:
      operationId: createPaymentRequest
      tags: [payments]
      summary: Submit a payment release request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequestBody"
      responses:
        "201":
          description: Payment request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
  /payment-requests/{id}:
    get:
      operationId: getPaymentRequest
      tags: [payments]
      summary: Get a payment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
    patch:
      operationId: updatePaymentRequest
      tags: [payments]
      summary: Approve/reject/release a payment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentRequestBody"
      responses:
        "200":
          description: Payment request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"

  # Tickets
...
components:
  schemas:
...
    Rider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        vehicleNumber:
          type: ["string", "null"]
        stationId:
          type: ["integer", "null"]
        stationName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        userId:
          type: ["integer", "null"]
        assignedCount:
          type: integer
        deliveredToday:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - status
        - assignedCount
        - deliveredToday
        - createdAt
    CreateRiderBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        vehicleNumber:
          type: ["string", "null"]
        stationId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
      required:
        - name
        - email
    UpdateRiderBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        vehicleNumber:
          type: ["string", "null"]
        stationId:
          type: ["integer", "null"]
        status:
          type: string
...
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        accountHolderName:
          type: string
        bankName:
          type: string
        branch:
          type: ["string", "null"]
        accountNumber:
          type: string
        walletMethod:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        isDefault:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - vendorId
        - accountHolderName
        - bankName
        - accountNumber
        - isDefault
        - createdAt
    CreateBankAccountBody:
      type: object
      properties:
        vendorId:
          type: integer
        accountHolderName:
          type: string
        bankName:
          type: string
        branch:
          type: ["string", "null"]
        accountNumber:
          type: string
        walletMethod:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        isDefault:
          type: boolean
      required:
        - vendorId
        - accountHolderName
        - bankName
        - accountNumber
    UpdateBankAccountBody:
      type: object
      properties:
        accountHolderName:
          type: string
        bankName:
          type: string
        branch:
          type: ["string", "null"]
        accountNumber:
          type: string
        walletMethod:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        isDefault:
          type: boolean
    PaymentRequest:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        bankAccountId:
          type: integer
        bankAccountInfo:
          type: string
        requestedAmount:
          type: number
        approvedAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, released]
        note:
          type: ["string", "null"]
        releaseNote:
          type: ["string", "null"]
        referenceId:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        reviewedByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - vendorId
        - vendorName
        - bankAccountId
        - bankAccountInfo
        - requestedAmount
        - status
        - createdAt
        - updatedAt
    CreatePaymentRequestBody:
      type: object
      properties:
        bankAccountId:
          type: integer
        requestedAmount:
          type: number
        note:
          type: ["string", "null"]
      required:
        - bankAccountId
        - requestedAmount
    UpdatePaymentRequestBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected, released]
        approvedAmount:
          type: ["number", "null"]
        releaseNote:
          type: ["string", "null"]
        referenceId:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
...